+++ Tests/HydraTests/HydraTests.swift
@@ -242,6 +242,20 @@ class HydraTestThen: XCTestCase {
 		waitForExpectations(timeout: expTimeout, handler: nil)
 	}
 	
+	/// If return rejected promise in `recover` operator, chain to next as its error.
+	func test_recover_failure() {
+		let exp = expectation(description: "test_recover_failure")
+		
+		let errPromise = intFailedPromiseImmediate(TestErrors.anotherError)
+		errPromise.recover { (err) -> Promise<Int> in
+			return Promise<Int>(rejected: TestErrors.someError)
+		}.catch { (e) -> (Void) in
+			XCTAssertEqual(e as! TestErrors, TestErrors.someError)
+			exp.fulfill()
+		}
+		waitForExpectations(timeout: expTimeout, handler: nil)
+	}
+	
 	//MARK: Pass tests
 	
 	
@@ -357,6 +371,32 @@ class HydraTestThen: XCTestCase {
 		waitForExpectations(timeout: expTimeout, handler: nil)
 	}
 	
+	func test_anyWithAllBodyPromise() {
+		let exp = expectation(description: "test_anyWithAllBodyPromise")
+		let promise1 = intPromiseDelay(100, delay: 0.1)
+		let promise2 = intPromise(5)
+		any(promise1, promise2).then { result in
+			XCTAssertEqual(result, promise1.result!)
+			exp.fulfill()
+			}.catch { _ in
+				XCTFail()
+		}
+		waitForExpectations(timeout: expTimeout, handler: nil)
+	}
+	
+	func test_anyWithArrayAllBodyPromise() {
+		let exp = expectation(description: "test_anyWithArrayAllBodyPromise")
+		let promise1 = intPromiseDelay(100, delay: 0.1)
+		let promise2 = intPromise(5)
+		any([promise1, promise2]).then { result in
+			XCTAssertEqual(result, promise1.result!)
+			exp.fulfill()
+			}.catch { _ in
+				XCTFail()
+		}
+		waitForExpectations(timeout: expTimeout, handler: nil)
+	}
+	
 	//MARK: All Tests
 	
 	
@@ -376,7 +416,25 @@ class HydraTestThen: XCTestCase {
 		}
 		waitForExpectations(timeout: expTimeout, handler: nil)
 	}
-
+	
+	/// `all` operator with comcurrency argument test.
+	/// body executable timing will restricted by concurrency.
+	func test_all_with_concurrency() {
+		let exp = expectation(description: "test_all_with_concurrency")
+		var timebasedResults = [Int]()
+		let promise1 = intPromiseDelayWithCompletion(3, delay: 0.5, completion: { timebasedResults.append($0) })
+		let promise2 = intPromiseDelayWithCompletion(12, delay: 0, completion: { timebasedResults.append($0) })
+		all(promise1, promise2, concurrency: 1).then { results in
+			XCTAssertEqual(results[0], 3)
+			XCTAssertEqual(results[1], 12)
+			XCTAssertEqual(timebasedResults[0], promise1.result!)
+			XCTAssertEqual(timebasedResults[1], promise2.result!)
+			exp.fulfill()
+		}.catch { _ in
+			XCTFail()
+		}
+		waitForExpectations(timeout: expTimeout, handler: nil)
+	}
 	
 	/// This is another test with `all` operator.
 	/// This test it's okay if all-promise is rejected because one of the input promise rejects.
@@ -593,6 +651,7 @@ class HydraTestThen: XCTestCase {
 			}
 		}.retry(retryAttempts).then { value in
 			print("value \(value) at attempt \(currentAttempt)")
+			XCTAssertEqual(currentAttempt, 3)
 			exp.fulfill()
 		}.catch { err in
 			print("failed \(err) at attempt \(currentAttempt)")
@@ -600,6 +659,29 @@ class HydraTestThen: XCTestCase {
 		}
 		
 		waitForExpectations(timeout: expTimeout, handler: nil)
+	}
+	
+	func test_retry_allFailure() {
+		let exp = expectation(description: "test_retry_allFailure")
+		
+		let retryAttempts = 3
+		var currentAttempt = 0
+		Promise<Int> { (resolve, reject) in
+			currentAttempt += 1
+			print("attempt is \(currentAttempt)... reject")
+			reject(TestErrors.anotherError)
+		}.retry(retryAttempts).then { value in
+			print("value \(value) at attempt \(currentAttempt)")
+			XCTFail()
+		}.catch { err in
+			print("failed \(err) at attempt \(currentAttempt)")
+			XCTAssertEqual(err as! TestErrors, .anotherError)
+			XCTAssertEqual(currentAttempt, 3)
+			exp.fulfill()
+		}
+		
+		waitForExpectations(timeout: expTimeout, handler: nil)
+		
 	}
 	
 	func test_retry_condition() {
@@ -631,7 +713,8 @@ class HydraTestThen: XCTestCase {
 			XCTFail()
 		}.catch { err in
 			print("failed \(err) at attempt \(currentAttempt)")
-    		XCTAssertEqual(currentAttempt, 3)
+			XCTAssertEqual(err as! TestErrors, .anotherError)
+			XCTAssertEqual(currentAttempt, 3)
 			exp.fulfill()
 		}
 		
@@ -671,6 +754,15 @@ class HydraTestThen: XCTestCase {
 	func intPromiseDelay(_ value: Int = 10, delay: TimeInterval) -> Promise<Int> {
 		return Promise<Int> { resolve, _ in
 			DispatchQueue.global(qos: .background).asyncAfter(deadline: .now() + delay, execute: {
+				resolve(value)
+			})
+		}
+	}
+	
+	func intPromiseDelayWithCompletion(_ value: Int = 10, delay: TimeInterval, completion: ((Int) -> Void)? = nil) -> Promise<Int> {
+		return Promise<Int> { resolve, _ in
+			DispatchQueue.global(qos: .background).asyncAfter(deadline: .now() + delay, execute: {
+				completion?(value)
 				resolve(value)
 			})
 		}