+++ CHANGELOG.md
@@ -4,9 +4,25 @@
 
 ## CHANGELOG
 
+* Version **[0.9.4](#094)**
 * Version **[0.9.3](#093)**
 * Version **[0.9.2](#092)**
 * Version **[0.9.1](#091)**
+
+<a name="094" />
+
+## Hydra 0.9.4
+---
+- **Release Date**: 2017-03-21
+- **Zipped Version**: [Download 0.9.4](https://github.com/malcommac/Hydra/releases/tag/0.9.4)
+
+- [#22](https://github.com/malcommac/Hydra/pull/22) Fixed an issue with `any` operator which in several occasions does not work.
+- [#24](https://github.com/malcommac/Hydra/pull/24) Fixed never chainable if return rejected promise in `recover` operator closure.
+- [#26](https://github.com/malcommac/Hydra/pull/26) Added concurrency argument in `all` operator; it allows you to set the number of max concurrent promises running for an all task in order to reduce memory and cpu footprint.
+- [#28](https://github.com/malcommac/Hydra/pull/28) Fixed Promise's memory leak
+- [#29](https://github.com/malcommac/Hydra/pull/29) Cleaned up the process of running the observer
+- [#18](https://github.com/malcommac/Hydra/pull/18), [#20](https://github.com/malcommac/Hydra/pull/20) Refactoring of several func inside the Promise's class.
+
 
 <a name="093" />